+++ astrcontinuum/adapters/__init__.py
@@ -20,9 +20,9 @@ from .astrbot import (
     extract_host_session_key,
     extract_host_turn_identity,
     projection_factory_from_user_message,
-    sanitize_claude_openai_contexts,
     select_projection_boundaries,
 )
+from .provider_thinking import ProviderThinkingCompatibility, ThinkingCompatibilityBlockedError
 from .sylanne import SylanneAdapter
 
 __all__ = [
@@ -38,7 +38,9 @@ __all__ = [
     "PreparedRequest",
     "ProjectionBuild",
     "ProjectionFactory",
+    "ProviderThinkingCompatibility",
     "SylanneAdapter",
+    "ThinkingCompatibilityBlockedError",
     "build_projection_objects",
     "canonical_tool_metadata",
     "deterministic_event_identity",
@@ -46,6 +48,5 @@ __all__ = [
     "extract_host_session_key",
     "extract_host_turn_identity",
     "projection_factory_from_user_message",
-    "sanitize_claude_openai_contexts",
     "select_projection_boundaries",
 ]